+++ cobs/util/misc.cpp
@@ -37,7 +37,7 @@ uint64_t get_phys_pages () {
 namespace cobs {
 
 uint64_t get_page_size() {
-    int page_size = getpagesize();
+    int64_t page_size = getpagesize();
     die_unless(page_size > 0);
     die_unless(page_size == 4096);     // todo check for experiments
     return (uint64_t)page_size;
